working multi-asset setup
