+++ assets/clp-s/output.json
@@ -1,0 +1,89 @@
+{
+  "mongod": {
+    "target_encoded_size=268435456": {
+      "ingest": {
+        "time_taken": 15.117038249969482,
+        "memory_average_B": 104170837.33333333,
+        "compressed_size": "17232340",
+        "decompressed_size": "3960196482",
+        "start_time": "2025-07-02 10:45:35"
+      },
+      "query_cold": [
+        {
+          "time_taken": 0.05772233009338379,
+          "memory_average_B": -1,
+          "result": "1620",
+          "start_time": "2025-07-02 10:45:50"
+        },
+        {
+          "time_taken": 0.7428240776062012,
+          "memory_average_B": -1,
+          "result": "13806",
+          "start_time": "2025-07-02 10:45:50"
+        },
+        {
+          "time_taken": 0.0562288761138916,
+          "memory_average_B": -1,
+          "result": "1",
+          "start_time": "2025-07-02 10:45:51"
+        },
+        {
+          "time_taken": 0.2189788818359375,
+          "memory_average_B": -1,
+          "result": "122",
+          "start_time": "2025-07-02 10:45:51"
+        },
+        {
+          "time_taken": 0.06542730331420898,
+          "memory_average_B": -1,
+          "result": "0",
+          "start_time": "2025-07-02 10:45:52"
+        },
+        {
+          "time_taken": 0.04697108268737793,
+          "memory_average_B": -1,
+          "result": "1620",
+          "start_time": "2025-07-02 10:45:52"
+        }
+      ],
+      "query_hot": [
+        {
+          "time_taken": 0.061243295669555664,
+          "memory_average_B": -1,
+          "result": "1620",
+          "start_time": "2025-07-02 10:45:52"
+        },
+        {
+          "time_taken": 0.7583272457122803,
+          "memory_average_B": 121151488.0,
+          "result": "13806",
+          "start_time": "2025-07-02 10:45:54"
+        },
+        {
+          "time_taken": 0.05251169204711914,
+          "memory_average_B": -1,
+          "result": "1",
+          "start_time": "2025-07-02 10:45:55"
+        },
+        {
+          "time_taken": 0.20410966873168945,
+          "memory_average_B": -1,
+          "result": "122",
+          "start_time": "2025-07-02 10:45:56"
+        },
+        {
+          "time_taken": 0.05556321144104004,
+          "memory_average_B": -1,
+          "result": "0",
+          "start_time": "2025-07-02 10:45:56"
+        },
+        {
+          "time_taken": 0.04632830619812012,
+          "memory_average_B": -1,
+          "result": "1620",
+          "start_time": "2025-07-02 10:45:57"
+        }
+      ]
+    }
+  }
+}
